Register more secure
It is really important that when a new account is created that no edge cases exist. There were issues before with invalid emails getting accepted and a user being able to not enter a first and last name. These have been addressed and fixed. Also the error messages are signifcantly better now and actually propagate correctly.